# libs/frontend/core-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cue the Music",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/albums": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "List Albums",
        "description": "List albums with optional search, genre, and decade filters.",
        "operationId": "list_albums_api_albums_get",
        "parameters": [
          {
            "name": "decade",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Decade"
            }
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Genre"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/albums/{album_id}": {
      "get": {
        "tags": [
          "albums"
        ],
        "summary": "Get Album",
        "description": "Retrieve a single album by ID.",
        "operationId": "get_album_api_albums__album_id__get",
        "parameters": [
          {
            "name": "album_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Album Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlbumGetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlbumGetResponse": {
        "properties": {
          "artist": {
            "type": "string",
            "title": "Artist"
          },
          "cover_art_thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Art Thumbnail Url"
          },
          "cover_art_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Art Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "discogs_release_id": {
            "type": "string",
            "title": "Discogs Release Id"
          },
          "genre_tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre Tags"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "style_tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style Tags"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "tracklist": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tracklist"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "type": "object",
        "required": [
          "artist",
          "cover_art_thumbnail_url",
          "cover_art_url",
          "created_at",
          "discogs_release_id",
          "genre_tags",
          "id",
          "label",
          "style_tags",
          "title",
          "tracklist",
          "updated_at",
          "year"
        ],
        "title": "AlbumGetResponse",
        "description": "Response schema for a single album with all fields."
      },
      "AlbumListResponse": {
        "properties": {
          "albums": {
            "items": {
              "$ref": "#/components/schemas/AlbumGetResponse"
            },
            "type": "array",
            "title": "Albums"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "albums",
          "count"
        ],
        "title": "AlbumListResponse",
        "description": "Response schema for a list of albums."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}